airflow/astro/requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://docs.astronomer.io/astro/runtime-image-architecture#provider-packages
duckdb==1.4.4
tabulate==0.9.0
pyarrow
paramiko
boto3
psycopg2-binary
databricks-sql-connector
databricks-sdk
